+++ RestApiTz/Controllers/ObjectInventoryController.cs
@@ -27,13 +27,15 @@ namespace RestApiTz.Controllers
         }
 
         [HttpPost]
-        public void Post([FromBody] JToken value)
+        public IActionResult Post([FromBody] JToken value)
         {
             var newObj = JsonConvert.DeserializeObject<ObjectInventoryView>(value.ToString());
 
             var mapper = new MapperConfiguration(cfg => cfg.CreateMap<ObjectInventoryView, ObjectInventoryDTO>()).CreateMapper();
             var mObject = mapper.Map<ObjectInventoryView, ObjectInventoryDTO>(newObj);
             objectService.AddObjectInventory(mObject);
+
+            return Ok();
         }
 
         [HttpGet]
@@ -43,9 +45,15 @@ namespace RestApiTz.Controllers
         }
 
         [HttpGet("{id}")]
-        public JArray Get(int id)
+        public IActionResult Get(int id)
         {
-            return JArray.Parse(JsonConvert.SerializeObject(objectService.GetObjects(id, null)));
+            if (objectService.GetObject(id) != null)
+            {
+                return new ObjectResult(JsonConvert.SerializeObject(objectService.GetObject(id)));
+            } else
+            {
+                return NotFound();
+            }
         }
 
         [Route("FilterSortPage/")]
@@ -53,11 +61,11 @@ namespace RestApiTz.Controllers
         public JArray Get([FromBody]JToken optionsFilter)
         {
             var dOptions = JsonConvert.DeserializeObject<Options>(optionsFilter.ToString());
-            return JArray.Parse(JsonConvert.SerializeObject(objectService.GetObjects(null, dOptions)));            
+            return JArray.Parse(JsonConvert.SerializeObject(objectService.GetObjects(dOptions)));            
         }
 
         [HttpPut]
-        public void Put([FromBody]JToken value)
+        public IActionResult Put([FromBody]JToken value)
         {
             var newType = JsonConvert.DeserializeObject<ObjectInventoryView>(value.ToString());
 
@@ -65,12 +73,22 @@ namespace RestApiTz.Controllers
             var mObject = mapper.Map<ObjectInventoryView, ObjectInventoryDTO>(newType);
 
             objectService.UpdateObject(mObject);
+
+            return Ok();
         }
 
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
-            objectService.RemoveObject(id);
+            if (objectService.GetObject(id) != null)
+            {
+                objectService.RemoveObject(id);
+
+                return NoContent();
+            } else
+            {
+                return NotFound();
+            }            
         }
     }
 }
